RUMM-987 Decouble reflection helpers to separate files
